requirements.txt
mosaicml>=0.12,<0.13
pre-commit>=2.18.1,<3
pytest>=7.2.1,<8
pytest_codeblocks>=0.16.1,<0.17
pytest-cov>=4,<5
pyright>=1.1.290,<1.2
toml>=0.10.2,<0.11
packaging>=21,<23
omegaconf>=2.2.3,<3
# datasets>=2.9.0
transformers>=4.11
mosaicml-streaming>=0.3.0
tqdm
tqdm-multiprocess
pybind11
zstandard
jsonlines